+++ serde-shape/src/impls/mod.rs
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod bound;
 mod container;
+mod ffi;
 mod net;
 mod primitive;
+mod range;
 mod result;
 mod time;
 mod tuple;